pyproject.toml
[project]
name = "who-reviews"
version = "0.1.0"
description = "GitHub Action for automated PR reviewer assignment based on code ownership"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[tool.uv]
dev-dependencies = [
    "mypy>=1.13",
    "pre-commit>=4.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
    "types-pyyaml>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_decorators = false
ignore_missing_imports = true
